+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `archivalist` will be documented in this file
 
 ### Added
 - Added support for JSON relationships, [see the repo for more details](https://github.com/staudenmeir/eloquent-json-relations)
+- Added BeforeArchive callback function. See readme for details on its usage.
 
 ## 0.2.2 - 2020-07-13